requirements.txt
openai-whisper==20231117
googletrans==3.1.0a0
